aggregation/examples: small refactor
